Reorganised TyrianApp and added Scaladocs
