+++ tyrian/js/src/main/scala/tyrian/TyrianAppF.scala
@@ -1,0 +1,62 @@
+package tyrian
+
+import cats.effect.kernel.Async
+import org.scalajs.dom.document
+import tyrian.runtime.RunWithCallback
+
+import scala.scalajs.js.annotation._
+
+/** The TyrianApp trait can be extended to conveniently prompt you for all the methods needed for a Tyrian app, as well
+  * as providing a number of standard app launching methods.
+  */
+trait TyrianAppF[F[_]: Async, Msg, Model]:
+
+  protected val runner: RunWithCallback[F, Msg]
+
+  /** Used to initialise your app. Accepts simple flags and produces the initial model state, along with any commands to
+    * run at start up, in order to trigger other processes.
+    */
+  def init(flags: Map[String, String]): (Model, Cmd[F, Msg])
+
+  /** The update method allows you to modify the model based on incoming messages (events). As well as an updated model,
+    * you can also produce commands to run.
+    */
+  def update(msg: Msg, model: Model): (Model, Cmd[F, Msg])
+
+  /** Used to render your current model into an HTML view.
+    */
+  def view(model: Model): Html[Msg]
+
+  /** Subscriptions are typically processes that run for a period of time and emit discrete events based on some world
+    * event, e.g. a mouse moving might emit it's coordinates.
+    */
+  def subscriptions(model: Model): Sub[F, Msg]
+
+  /** Launch the app and attach it to an element with the given id. Can be called from Scala or JavaScript.
+    */
+  @JSExport
+  def launch(containerId: String): Unit =
+    ready(containerId, Map[String, String]())
+
+  /** Launch the app and attach it to an element with the given id, with the supplied simple flags. Can be called from
+    * Scala or JavaScript.
+    */
+  @JSExport
+  def launch(containerId: String, flags: scala.scalajs.js.Dictionary[String]): Unit =
+    ready(containerId, flags.toMap)
+
+  /** Launch the app and attach it to an element with the given id, with the supplied simple flags. Can only be called
+    * from Scala.
+    */
+  def launch(containerId: String, flags: Map[String, String]): Unit =
+    ready(containerId, flags)
+
+  private def ready(parentElementId: String, flags: Map[String, String]): Unit =
+    Tyrian.start(
+      document.getElementById(parentElementId),
+      init(flags),
+      update,
+      view,
+      subscriptions,
+      runner
+    )
